Update yes.py
When I made this terminal quiz, I was new to python, so I cam back and fixed something that bugged me. In the line where it says "welcome to my sports quiz", it used to say type everything in lower case. I cam back, deleted that, and added .lower() to the inputs. I don't think this will be my last change though.